refactor: update data types for calories and total calories in response DTOs, and enhance SQL schema for roles. Added Reports controller, services + covered with integration tests

+++ src/main/java/source/code/dto/response/activity/ActivityCalculatedResponseDto.java
@@ -25,7 +25,7 @@ public class ActivityCalculatedResponseDto implements Serializable {
 
 	private CategoryResponseDto category;
 
-	private int caloriesBurned;
+	private BigDecimal caloriesBurned;
 
 	private int time;
